+++ backend/fundraising-service/src/main/java/com/alumniconnect/fundraising/repository/DonationRepository.java
@@ -7,7 +7,7 @@ import java.util.List;
 
 @Repository
 public interface DonationRepository extends JpaRepository<Donation, Long> {
-    List<Donation> findByFundraisingFundId(Integer fundId);
+    List<Donation> findByFundraisingFundId(Long fundId);
     List<Donation> findByAlumniId(Integer alumniId);
     List<Donation> findByPaymentStatus(String paymentStatus);
 }
