+++ backend/fundraising-service/src/main/java/com/alumniconnect/fundraising/service/DonationService.java
@@ -5,7 +5,7 @@ import com.alumniconnect.fundraising.entity.Donation;
 
 public interface DonationService {
     Donation processDonation(Donation donation);
-    List<Donation> getDonationsByCampaign(Integer fundId);
+    List<Donation> getDonationsByCampaign(Long fundId);
     List<Donation> getDonationsByAlumni(Integer alumniId);
     List<Donation> getAllDonations();
 }
