+++ backend/mentorship-service/src/main/java/com/alumniconnect/mentorship/service/MentorshipService.java
@@ -6,7 +6,7 @@ import com.alumniconnect.mentorship.entity.MentorshipRequest;
 public interface MentorshipService {
     MentorshipRequest addMentorship(MentorshipRequest mentorship);
     MentorshipRequest updateMentorship(MentorshipRequest mentorship);
-    void deleteMentorship(Integer requestId);
-    MentorshipRequest getMentorshipById(Integer requestId);
+    void deleteMentorship(Long requestId);
+    MentorshipRequest getMentorshipById(Long requestId);
     List<MentorshipRequest> getAllMentorships();
 }
